+++ types/index.d.ts
@@ -1,1 +1,2 @@
+import './node_windows_x64/define';
 export * as node_windows_x64 from "./node_windows_x64";