+++ types/node_windows_x64/win32_gui.d.ts
@@ -1,0 +1,70 @@
+export interface Win32CreateOption {
+  id: number;
+  windowName?: string;
+  style?: number;
+  x?: number;
+  y?: number;
+  width?: number;
+  height?: number;
+  parent?: number;
+  events?: {
+    [type: string]: Function;
+  };
+}
+
+export class Win32Gui {
+  static getCheck(hwnd: number): boolean;
+  static rgb(r: number, g: number, b: number): number;
+  constructor(
+    className: string,
+    windowName: string,
+    opt?: {
+      x?: number;
+      y?: number;
+      width?: number;
+      height?: number;
+      style?: number;
+    }
+  );
+
+  initRegisterClass(): boolean;
+  initWindow(): boolean;
+  messageLoop(
+    OnReceiveMessage?: (
+      hWnd: number,
+      message: number,
+      wParam: number,
+      lParam: number
+    ) => undefined | number
+  ): number;
+
+  createWindow(opt: {
+    className: string;
+    id: number;
+    windowName?: string;
+    style?: number;
+    x?: number;
+    y?: number;
+    width?: number;
+    height?: number;
+    parent?: number;
+  }): number;
+
+  button(opt: Win32CreateOption): number;
+
+  checkbox(opt: Win32CreateOption): number;
+
+  radio(opt: Win32CreateOption): number;
+
+  groupbox(opt: Win32CreateOption): number;
+
+  text(opt: Win32CreateOption): number;
+
+  input(opt: Win32CreateOption): number;
+
+  textarea(opt: Win32CreateOption): number;
+
+  listbox(opt: Win32CreateOption): number;
+
+  select(opt: Win32CreateOption): number;
+}